Break up the Transfer files
